Move Expression to showA
