+++ src/Ampersand/FSpec/ShowADL.hs
@@ -18,7 +18,7 @@ import Ampersand.Core.ParseTree
      , Prop(..)
      )
 import Ampersand.Core.ShowPStruct
---import Ampersand.Core.ShowAStruct
+-- import Ampersand.Core.ShowAStruct(showA)
 import Ampersand.Core.AbstractSyntaxTree
      ( Cruds(..)
      , Meta(..)
@@ -47,7 +47,6 @@ import Ampersand.Core.AbstractSyntaxTree
      )
 import Ampersand.Basics hiding (indent)
 import Ampersand.Classes
-import Ampersand.ADL1 (insParentheses)
 import Ampersand.FSpec.FSpec
 import Data.List
 import Prelude
@@ -205,43 +204,8 @@ instance ShowADL ViewSegmentPayLoad where
 -- showADL rel = show rel
 
 instance ShowADL Expression where
- showADL = showExpr (" = ", " |- ", " /\\ ", " \\/ ", " - ", " / ", " \\ ", " <> ", ";", "!", "*", "*", "+", "~", ("-"++), "(", ")", "[", "*", "]")
--- NOTE: retain space after \\, because of unexpected side effects if it is used just before an 'r' or 'n'....
-   where
-     showExpr :: (String,String,String,String,String,String,String,String,String,String,String,String,String,String,String -> String,String,String,String,String,String)
-            -> Expression -> String
-     showExpr    (equiv,  inclu,  inter, union',diff,  lresi, rresi, rDia, rMul  , rAdd , rPrd ,closK0,closK1,flp',  compl,           lpar,  rpar,  lbr,   star,  rbr)  expr
-      = showchar (insParentheses expr)
-        where
-          showchar (EEqu (l,r)) = showchar l++equiv++showchar r
-          showchar (EInc (l,r)) = showchar l++inclu++showchar r
-          showchar (EIsc (l,r)) = showchar l++inter++showchar r
-          showchar (EUni (l,r)) = showchar l++union'++showchar r
-          showchar (EDif (l,r)) = showchar l++diff ++showchar r
-          showchar (ELrs (l,r)) = showchar l++lresi++showchar r
-          showchar (ERrs (l,r)) = showchar l++rresi++showchar r
-          showchar (EDia (l,r)) = showchar l++rDia++showchar r
-          showchar (ECps (l,r)) = showchar l++rMul++showchar r
-          showchar (ERad (l,r)) = showchar l++rAdd++showchar r
-          showchar (EPrd (l,r)) = showchar l++rPrd++showchar r
-          showchar (EKl0 e)     = showchar e++closK0
-          showchar (EKl1 e)     = showchar e++closK1
-          showchar (EFlp e)     = showchar e++flp'
-          showchar (ECpl e)     = compl (showchar e)
-          showchar (EBrk e)     = lpar++showchar e++rpar
-          showchar (EDcD dcl)   = name dcl
-          showchar (EDcI c)     = "I"++lbr++name c++rbr
-          showchar (EEps i _)   = "I{-Eps-}"++lbr++name i++rbr
-          showchar (EDcV sgn)   = "V"++lbr++name (source sgn)++star++name (target sgn)++rbr
-          showchar (EMp1 val c) = "'"++showWithoutDoubleQuotes val++"'"++lbr++name c++rbr
-            
-          showWithoutDoubleQuotes str = 
-            case showADL str of
-              []  -> []
-              [c] -> [c]
-              cs  -> if head cs == '\"' && last cs == '\"'
-                     then reverse . tail . reverse .tail $ cs
-                     else cs
+ showADL = showADL
+
 instance ShowADL DnfClause where
  showADL dnfClause = showADL (dnf2expr dnfClause)
